Adding the version of the specification used to create a workflow to the workflow api endpoint.  Though the exact content of this version is likely to change. Split the API specific models out from the workflow models to help me keep this straight. Added tests to help me understand the errors thrown the and resolution path when a workflow specification changes in the midst of a running workflow.
