+++ tests/test_workflow_processor.py
@@ -7,7 +7,7 @@ from SpiffWorkflow.bpmn.specs.EndEvent import EndEvent
 
 from crc import session, db, app
 from crc.api.common import ApiError
-from crc.models.file import FileModel, FileDataModel
+from crc.models.file import FileModel, FileDataModel, CONTENT_TYPES
 from crc.models.study import StudyModel
 from crc.models.workflow import WorkflowSpecModel, WorkflowStatus, WorkflowModel
 from crc.services.file_service import FileService
@@ -254,3 +254,66 @@ class TestWorkflowProcessor(BaseTest):
         self.assertNotEqual(task.get_name(), task_before_restart.get_name())
         self.assertEqual(task.get_name(), task_after_restart.get_name())
         self.assertEqual(task.data, task_after_restart.data)
+
+    def replace_file(self, name, file_path):
+        """Replaces a stored file with the given name with the contents of the file at the given path."""
+        file_service = FileService()
+        file = open(file_path, "rb")
+        data = file.read()
+
+        file_model = db.session.query(FileModel).filter(FileModel.name == name).first()
+        noise, file_extension = os.path.splitext(file_path)
+        content_type = CONTENT_TYPES[file_extension[1:]]
+        file_service.update_file(file_model, data, content_type)
+
+    def test_modify_spec_with_text_change_with_running_workflow(self):
+        self.load_example_data()
+
+        # Start the two_forms workflow, and enter some data in the first form.
+        study = session.query(StudyModel).first()
+        workflow_spec_model = self.load_test_spec("two_forms")
+        processor = WorkflowProcessor.create(study.id, workflow_spec_model.id)
+        workflow_model = db.session.query(WorkflowModel).filter(WorkflowModel.study_id == study.id).first()
+        self.assertEqual(workflow_model.workflow_spec_id, workflow_spec_model.id)
+        task = processor.next_task()
+        task.data = {"color": "blue"}
+        processor.complete_task(task)
+
+        # Modify the specification, with a minor text change.
+        file_path = os.path.join(app.root_path, '..', 'tests', 'data', 'two_forms', 'mods', 'two_forms_text_mod.bpmn')
+        self.replace_file("two_forms.bpmn", file_path)
+
+        processor2 = WorkflowProcessor(workflow_spec_model.id, processor.serialize())
+        self.assertEquals("Step 1", processor2.bpmn_workflow.last_task.task_spec.description)
+        self.assertEquals("# This is some documentation I wanted to add.",
+                          processor2.bpmn_workflow.last_task.task_spec.documentation)
+
+
+    def test_modify_spec_with_structural_change_with_running_workflow(self):
+        self.load_example_data()
+
+        # Start the two_forms workflow, and enter some data in the first form.
+        study = session.query(StudyModel).first()
+        workflow_spec_model = self.load_test_spec("two_forms")
+        processor = WorkflowProcessor.create(study.id, workflow_spec_model.id)
+        workflow_model = db.session.query(WorkflowModel).filter(WorkflowModel.study_id == study.id).first()
+        self.assertEqual(workflow_model.workflow_spec_id, workflow_spec_model.id)
+        task = processor.next_task()
+        task.data = {"color": "blue"}
+        processor.complete_task(task)
+        next_task = processor.next_task()
+        self.assertEquals("Step 2", next_task.task_spec.description)
+
+        # Modify the specification, with a major change that alters the flow and can't be serialized effectively.
+        file_path = os.path.join(app.root_path, '..', 'tests', 'data', 'two_forms', 'mods', 'two_forms_struc_mod.bpmn')
+        self.replace_file("two_forms.bpmn", file_path)
+
+        with self.assertRaises(KeyError):
+            processor2 = WorkflowProcessor(workflow_spec_model.id, processor.serialize())
+
+        # Restart the workflow, and the error should go away
+        processor.restart_with_current_task_data()
+        self.assertEquals("Step 1", processor.next_task().task_spec.description)
+        processor.complete_task(processor.next_task())
+        self.assertEquals("New Step", processor.next_task().task_spec.description)
+        self.assertEquals({"color": "blue"}, processor.next_task().data)
